Funzioni.createM2 — Function

createM2(l, t)

Create an M2 matrix that has the faces on the rows and the vertices on the columns.

Arguments:

l: the faces list of object

t: the number of vertices.

Examples:

julia> createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)
6×8 Array{Int64,2}:
    1  0  1  0  1  0  1  0
    0  0  1  1  0  0  1  1
    0  0  0  0  1  1  1  1
    0  1  0  1  0  1  0  1
    1  1  1  1  0  0  0  0
    1  1  0  0  1  1  0  0

Funzioni.createM1 — Function

createM1(m)

Create an M1 matrix that has the edges on the rows and the vertices on the columns.

Arguments:

m: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> createM1(matrixM2)
12×8 Array{Int64,2}:
     0  0  1  0  0  0  1  0
     0  0  0  0  1  0  1  0
     1  0  1  0  0  0  0  0
     1  0  0  0  1  0  0  0
     0  0  0  0  0  0  1  1
     0  0  0  1  0  0  0  1
     0  0  1  1  0  0  0  0
     0  0  0  0  0  1  0  1
     0  0  0  0  1  1  0  0
     0  1  0  1  0  0  0  0
     0  1  0  0  0  1  0  0
     1  1  0  0  0  0  0  0

Funzioni.getMatrixVV — Function

getMatrixVV(m)

Create a VV GBMatrix that has the vertices on the rows and the vertex on the columns from an M1 matrix.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixVV(matrixM1) 
8x8 GBMatrix{Int64} with 32 stored entries:
  [1, 1] = 1
  [1, 2] = 0
  [1, 3] = 0
  [1, 4] = 0
  [1, 5] = 0
  [1, 6] = 0
  [1, 7] = 0
  [1, 8] = 0
  [2, 1] = 0
  [2, 2] = 1
  [2, 3] = 0
  [2, 4] = 0
  [2, 5] = 0
  [2, 6] = 0
  [2, 7] = 0
  [2, 8] = 0
  ⋮

  [6, 8] = 0
  [7, 1] = 0
  [7, 2] = 0
  [7, 3] = 0
  [7, 4] = 0
  [7, 5] = 0
  [7, 6] = 0
  [7, 7] = 1
  [7, 8] = 0
  [8, 1] = 0
  [8, 2] = 0
  [8, 3] = 0
  [8, 4] = 0
  [8, 5] = 0
  [8, 6] = 0
  [8, 7] = 0
  [8, 8] = 1

Funzioni.getMatrixVE — Function

getMatrixVE(m)

Create a VE GBMatrix that has the vertices on the rows and the edges on the columns from an M1 matrix.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixVE(matrixM1) 
8x12 GBMatrix{Int64} with 24 stored entries:
  [1, 3] = 1
  [1, 4] = 1
  [1, 12] = 1
  [2, 10] = 1
  [2, 11] = 1
  [2, 12] = 1
  [3, 1] = 1
  [3, 3] = 1
  [3, 7] = 1
  [4, 6] = 1
  [4, 7] = 1
  [4, 10] = 1
  [5, 2] = 1
  [5, 4] = 1
  [5, 9] = 1
  [6, 8] = 1
  [6, 9] = 1
  [6, 11] = 1
  [7, 1] = 1
  [7, 2] = 1
  [7, 5] = 1
  [8, 5] = 1
  [8, 6] = 1
  [8, 8] = 1

Funzioni.getMatrixVF — Function

getMatrixVF(m)

Create a VF GBMatrix that has the vertices on the rows and the faces on the columns from an M2 matrix.

Arguments:

m: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> getMatrixVF(matrixM2)
8x6 GBMatrix{Int64} with 24 stored entries:
  [1, 1] = 1
  [1, 5] = 1
  [1, 6] = 1
  [2, 4] = 1
  [2, 5] = 1
  [2, 6] = 1
  [3, 1] = 1
  [3, 2] = 1
  [3, 5] = 1
  [4, 2] = 1
  [4, 4] = 1
  [4, 5] = 1
  [5, 1] = 1
  [5, 3] = 1
  [5, 6] = 1
  [6, 3] = 1
  [6, 4] = 1
  [6, 6] = 1
  [7, 1] = 1
  [7, 2] = 1
  [7, 3] = 1
  [8, 2] = 1
  [8, 3] = 1
  [8, 4] = 1

Funzioni.getMatrixEV — Function

getMatrixEV(m)

Create a EV GBMatrix that has the edges on the rows and the vertices on the columns from an M1 matrix.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixEV(matrixM1)
12x8 GBMatrix{Int64} with 24 stored entries:
  [ 1, 3] = 1
  [ 1, 7] = 1
  [ 2, 5] = 1
  [ 2, 7] = 1
  [ 3, 1] = 1
  [ 3, 3] = 1
  [ 4, 1] = 1
  [ 4, 5] = 1
  [ 5, 7] = 1
  [ 5, 8] = 1
  [ 6, 4] = 1
  [ 6, 8] = 1
  [ 7, 3] = 1
  [ 7, 4] = 1
  [ 8, 6] = 1
  [ 8, 8] = 1
  [ 9, 5] = 1
  [ 9, 6] = 1
  [10, 2] = 1
  [10, 4] = 1
  [11, 2] = 1
  [11, 6] = 1
  [12, 1] = 1
  [12, 2] = 1

Funzioni.getMatrixEE — Function

getMatrixEE(m)

Create a EE GBMatrix that has the edges on the rows and columns from an M1 matrix.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixEE(matrixM1)
12x12 GBMatrix{Int64} with 60 stored entries:
  [ 1,  1] = 1
  [ 1,  2] = 0
  [ 1,  3] = 0
  [ 1,  4] = 0
  [ 1,  5] = 0
  [ 1,  6] = 0
  [ 1,  7] = 0
  [ 1,  8] = 0
  [ 1,  9] = 0
  [ 1, 10] = 0
  [ 1, 11] = 0
  [ 1, 12] = 0
  [ 2,  1] = 0
  [ 2,  2] = 1
  [ 2,  3] = 0
  [ 2,  4] = 0
  ⋮

  [11,  8] = 0
  [11,  9] = 0
  [11, 10] = 0
  [11, 11] = 1
  [11, 12] = 0
  [12,  1] = 0
  [12,  2] = 0
  [12,  3] = 0
  [12,  4] = 0
  [12,  5] = 0
  [12,  6] = 0
  [12,  7] = 0
  [12,  8] = 0
  [12,  9] = 0
  [12, 10] = 0
  [12, 11] = 0
  [12, 12] = 1

Funzioni.getMatrixEF — Function

getMatrixEF(m,n)

Create a EF GBMatrix that has the edges on the rows and the faces on the columns from M1 and M2 matrices.

Arguments:

m: the M1 matrix of edges

n: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixEF(matrixM1,matrixM2)
12x6 GBMatrix{Int64} with 48 stored entries:
  [ 1, 1] = 1
  [ 1, 2] = 1
  [ 1, 3] = 0
  [ 1, 4] = 0
  [ 1, 5] = 0
  [ 1, 6] = 0
  [ 2, 1] = 1
  [ 2, 2] = 0
  [ 2, 3] = 1
  [ 2, 4] = 0
  [ 2, 5] = 0
  [ 2, 6] = 0
  [ 3, 1] = 1
  [ 3, 2] = 0
  [ 3, 3] = 0
  [ 3, 4] = 0
  ⋮

  [10, 2] = 0
  [10, 3] = 0
  [10, 4] = 1
  [10, 5] = 1
  [10, 6] = 0
  [11, 1] = 0
  [11, 2] = 0
  [11, 3] = 0
  [11, 4] = 1
  [11, 5] = 0
  [11, 6] = 1
  [12, 1] = 0
  [12, 2] = 0
  [12, 3] = 0
  [12, 4] = 0
  [12, 5] = 1
  [12, 6] = 1

Funzioni.getMatrixFV — Function

getMatrixFV(m)

Create a FV GBMatrix that has the faces on the rows and the vertices on the columns from an M2 matrix.

Arguments:

m: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> getMatrixFV(matrixM2)
6x8 GBMatrix{Int64} with 24 stored entries:
  [1, 1] = 1
  [1, 3] = 1
  [1, 5] = 1
  [1, 7] = 1
  [2, 3] = 1
  [2, 4] = 1
  [2, 7] = 1
  [2, 8] = 1
  [3, 5] = 1
  [3, 6] = 1
  [3, 7] = 1
  [3, 8] = 1
  [4, 2] = 1
  [4, 4] = 1
  [4, 6] = 1
  [4, 8] = 1
  [5, 1] = 1
  [5, 2] = 1
  [5, 3] = 1
  [5, 4] = 1
  [6, 1] = 1
  [6, 2] = 1
  [6, 5] = 1
  [6, 6] = 1

Funzioni.getMatrixFE — Function

getMatrixFE(m,n)

Create a FE GBMatrix that has the faces on the rows and the edges on the columns from M1 and M2 matrices.

Arguments:

m: the M1 matrix of edges

n: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixFE(matrixM1,matrixM2)
6x12 GBMatrix{Int64} with 48 stored entries:
  [1,  1] = 0
  [1,  2] = 0
  [1,  3] = 0
  [1,  4] = 0
  [1,  5] = 0
  [1,  6] = 0
  [1,  7] = 0
  [1,  8] = 0
  [1,  9] = 0
  [1, 10] = 0
  [1, 11] = 0
  [1, 12] = 0
  [2,  1] = 0
  [2,  2] = 0
  [2,  3] = 0
  [2,  4] = 0
  ⋮

  [5,  8] = 0
  [5,  9] = 0
  [5, 10] = 0
  [5, 11] = 0
  [5, 12] = 0
  [6,  1] = 0
  [6,  2] = 0
  [6,  3] = 0
  [6,  4] = 0
  [6,  5] = 0
  [6,  6] = 0
  [6,  7] = 0
  [6,  8] = 0
  [6,  9] = 0
  [6, 10] = 0
  [6, 11] = 0
  [6, 12] = 0

Funzioni.getMatrixFF — Function

getMatrixFF(m)

Create a FF GBMatrix that has the faces on the rows and the columns from an M2 matrix.

Arguments:

m: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> getMatrixFF(matrixM2)
6x6 GBMatrix{Int64} with 30 stored entries:
  [1, 1] = 1
  [1, 2] = 0
  [1, 3] = 0
  [1, 4] = 0
  [1, 5] = 0
  [1, 6] = 0
  [2, 1] = 0
  [2, 2] = 1
  [2, 3] = 0
  [2, 4] = 0
  [2, 5] = 0
  [2, 6] = 0
  [3, 1] = 0
  [3, 2] = 0
  [3, 3] = 1
  [3, 4] = 0
  ⋮

  [4, 2] = 0
  [4, 3] = 0
  [4, 4] = 1
  [4, 5] = 0
  [4, 6] = 0
  [5, 1] = 0
  [5, 2] = 0
  [5, 3] = 0
  [5, 4] = 0
  [5, 5] = 1
  [5, 6] = 0
  [6, 1] = 0
  [6, 2] = 0
  [6, 3] = 0
  [6, 4] = 0
  [6, 5] = 0
  [6, 6] = 1