Funzioni.createM2 — Function

createM2(l, t)

Create an M2 matrix that has the faces on the rows and the vertex on the columns.

Arguments:

l: the faces list of object

t: the number of vertex.

Examples:

julia> createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)
6×8 Array{Int64,2}:
    1  0  1  0  1  0  1  0
    0  0  1  1  0  0  1  1
    0  0  0  0  1  1  1  1
    0  1  0  1  0  1  0  1
    1  1  1  1  0  0  0  0
    1  1  0  0  1  1  0  0

Funzioni.createM1 — Function

createM1(m)

Create an M1 matrix that has the edges on the rows and the vertex on the columns.

Arguments:

m: the M2 matrix of faces

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> createM1(matrixM2)
12×8 Array{Int64,2}:
     0  0  1  0  0  0  1  0
     0  0  0  0  1  0  1  0
     1  0  1  0  0  0  0  0
     1  0  0  0  1  0  0  0
     0  0  0  0  0  0  1  1
     0  0  0  1  0  0  0  1
     0  0  1  1  0  0  0  0
     0  0  0  0  0  1  0  1
     0  0  0  0  1  1  0  0
     0  1  0  1  0  0  0  0
     0  1  0  0  0  1  0  0
     1  1  0  0  0  0  0  0

Funzioni.getMatrixVV — Function

getMatrixVV(m)

Create a VV GBMatrix from an M1 matrix that has the vertex on the rows and the vertex on the columns.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixVV(matrixM1) 
8x8 GBMatrix{Int64} with 32 stored entries:
  [1, 1] = 3
  [1, 2] = 1
  [1, 3] = 1
  [1, 5] = 1
  [2, 1] = 1
  [2, 2] = 3
  [2, 4] = 1
  [2, 6] = 1
  [3, 1] = 1
  ⋮
  [6, 6] = 3
  [6, 8] = 1
  [7, 3] = 1
  [7, 5] = 1
  [7, 7] = 3
  [7, 8] = 1
  [8, 4] = 1
  [8, 6] = 1
  [8, 7] = 1
  [8, 8] = 3

Funzioni.getMatrixVE — Function

getMatrixVE(M1)

Create a VE GBMatrix from an M1 matrix that has the vertex on the rows and the edges on the columns.

Arguments:

m: the M1 matrix of edges

Examples:

julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixVE(matrixM1) 
8x12 GBMatrix{Int64} with 24 stored entries:
  [1, 3] = 1
  [1, 4] = 1
  [1, 12] = 1
  [2, 10] = 1
  [2, 11] = 1
  [2, 12] = 1
  [3, 1] = 1
  [3, 3] = 1
  [3, 7] = 1
  ⋮
  [5, 9] = 1
  [6, 8] = 1
  [6, 9] = 1
  [6, 11] = 1
  [7, 1] = 1
  [7, 2] = 1
  [7, 5] = 1
  [8, 5] = 1
  [8, 6] = 1
  [8, 8] = 1

Funzioni.getMatrixVF — Function

getMatrixVF(M2)

Create a VF GBMatrix from an M2 matrix that has the vertex on the rows and the faces on the columns.

Arguments:

m: the M2 matrix of faces

Examples:

```julia-repl julia> matrixM2 = createM2([[1, 5, 7, 3], [4, 3, 7, 8], [8, 7, 5, 6],[6, 2, 4, 8],[2, 1, 3, 4],[6, 5, 1, 2]], 8)

julia> matrixM1 = createM1(matrixM2)

julia> getMatrixVF(matrixM2) 8x6 GBMatrix{Int64} with 24 stored entries: [1, 1] = 1 [1, 5] = 1 [1, 6] = 1 [2, 4] = 1 [2, 5] = 1 [2, 6] = 1 [3, 1] = 1 [3, 2] = 1 [3, 5] = 1 ⋮ [5, 6] = 1 [6, 3] = 1 [6, 4] = 1 [6, 6] = 1 [7, 1] = 1 [7, 2] = 1 [7, 3] = 1 [8, 2] = 1 [8, 3] = 1 [8, 4] = 1

Funzioni.getMatrixEV — Function

getMatrixEV(M1) 
getMatrixEV

Funzioni.getMatrixEE — Function

getMatrixEE(M1)
getMatrixEE

Funzioni.getMatrixEF — Function

getMatrixEF(M1,M2)
getMatrixEF

Funzioni.getMatrixFV — Function

getMatrixFV(M2)
getMatrixFV

Funzioni.getMatrixFE — Function

getMatrixFE(M1,M2)
getMatrixFE

Funzioni.getMatrixFF — Function

getMatrixFF(M2)
getMatrixFF